# openapi.yaml
openapi: 3.0.3
info:
  title: AidLink Backend API
  version: 1.0.0
  description: |
    Production-grade REST API for AidLink — a blockchain-powered humanitarian aid platform built on Soroban/Stellar.

    ## Authentication
    All protected endpoints require a Bearer JWT in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
    Tokens are obtained via `POST /api/v1/auth/login` or `POST /api/v1/auth/wallet`.
    Access tokens expire in 15 minutes; use `POST /api/v1/auth/refresh` with the refresh token.

    ## Rate Limits
    Rate limit headers are returned on every response (`RateLimit-Limit`, `RateLimit-Remaining`, `RateLimit-Reset`).
    | Scope | Limit |
    |---|---|
    | Global API | 100 req / 15 min per IP |
    | Auth endpoints | 5 req / 15 min per IP |
    | Donation creation | 10 req / min per user |
    | Campaign creation | 3 req / hour per user |
    | Distribution creation | 5 req / min per user |
    | Search | 30 req / min per user |
    | Analytics | 20 req / min per user |
    | Notifications | 10 req / min per user |
    | Fraud reports | 5 req / hour per user |

    ## Roles
    | Role | Description |
    |---|---|
    | ADMIN | Full platform access |
    | ORGANIZATION | Manages campaigns and distributions |
    | DONOR | Creates donations |
    | BENEFICIARY | Receives distributions, submits KYC |
    | VERIFIER | Reviews KYC and beneficiary data |
    | AUDITOR | Read-only audit access |

    ## Errors
...
    ```json
...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/wallet
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - EMAIL_NOT_VERIFIED
                - VERIFICATION_FAILED
                - NOT_FOUND
                - GONE
                - CONFLICT
                - BAD_REQUEST
                - PAYLOAD_TOO_LARGE
                - UNSUPPORTED_MEDIA_TYPE
                - RATE_LIMITED
                - INTERNAL_SERVER_ERROR
              example: VALIDATION_ERROR
            errorCode:
              type: string
              description: >
                Stable domain-specific taxonomy code (e.g. "CAMPAIGN_002"), present on
                business-logic errors. See docs/ERROR_CODES.md for the full catalog with HTTP
                status, cause, and recommended resolution for every code.
              example: CAMPAIGN_002
            message:
              type: string
              example: Validation failed
            details:
              type: object
              additionalProperties:
                type: string
              example:
                amount: Amount must be positive

    # ── Enums ──────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication token required
    Forbidden:
      description: Authenticated but insufficient role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Insufficient permissions
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
    ValidationError:
      description: Request body failed schema validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                amount: Amount must be positive
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: RATE_LIMITED
              message: Too many requests, please try again later
...
paths:
...
  /api/v1/analytics/campaign/{campaignId}:
    get:
      tags: [Analytics]
      summary: Get campaign analytics
      description: |
        Returns detailed metrics for a campaign including donation trends, donor counts,
        matching stats, and milestone progress.
        **Auth:** Bearer JWT — campaign owner, ORGANIZATION, or ADMIN.
        **Rate limit:** 20 req / min per user.
      operationId: getCampaignAnalytics
      security:
        - bearerAuth: []
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
          example: clxcamp001
      responses:
        "200":
          description: Campaign analytics retrieved
          content:
            application/json:
              example:
                success: true
                data:
                  campaignId: clxcamp001
                  totalDonations: 47
                  totalDonated: "12500.00"
                  totalMatched: "3750.00"
                  uniqueDonors: 41
                  averageDonation: "265.96"
                  donationsByDay:
                    - date: "2024-03-15"
                      amount: "1250.00"
                      count: 5
                  milestoneProgress:
                    - id: clxmile001
                      title: Drill first borehole
                      targetAmount: "10000.00"
                      achieved: true
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "429":
          $ref: '#/components/responses/RateLimited'
  /api/v1/analytics/donor:
    get:
      tags: [Analytics]
      summary: Get donor analytics
      description: |
        Returns analytics for the authenticated donor.
        **Auth:** Bearer JWT — role: DONOR or ADMIN.
        **Rate limit:** 20 req / min per user.
      operationId: getDonorAnalytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Donor analytics retrieved
          content:
            application/json:
              example:
                success: true
                data:
                  totalDonated: "1250.00"
                  donationCount: 7
                  campaignsSupported: 4
                  averageDonation: "178.57"
                  donationsByMonth:
                    - month: "2024-03"
                      amount: "500.00"
                      count: 2
        "401":
          $ref: '#/components/responses/Unauthorized'
        "429":
          $ref: '#/components/responses/RateLimited'
  /api/v1/analytics/organization/{organizationId}:
    get:
      tags: [Analytics]
      summary: Get organization analytics
      description: |
        Returns aggregated analytics across all campaigns of an organization.
        **Auth:** Bearer JWT — organization owner or ADMIN.
        **Rate limit:** 20 req / min per user.
      operationId: getOrganizationAnalytics
      security:
        - bearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Organization analytics retrieved
          content:
            application/json:
              example:
                success: true
                data:
                  organizationId: clx2org0001
                  activeCampaigns: 3
                  totalCampaigns: 8
                  totalRaised: "245000.00"
                  totalDistributed: "198000.00"
                  beneficiariesReached: 142
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "429":
          $ref: '#/components/responses/RateLimited'
  /api/v1/analytics/platform:
    get:
      tags: [Analytics]
      summary: Get platform-wide analytics
      description: |
        **Auth:** Bearer JWT — role: ADMIN or AUDITOR.
        **Rate limit:** 20 req / min per user.
      operationId: getPlatformAnalytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platform analytics retrieved
          content:
            application/json:
              example:
                success: true
                data:
                  totalUsers: 1240
                  totalOrganizations: 38
                  totalCampaigns: 156
                  activeCampaigns: 34
                  totalDonated: "2345000.00"
                  totalDistributed: "1890000.00"
                  beneficiariesReached: 3420
                  blockchainTransactions: 5678
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "429":
          $ref: '#/components/responses/RateLimited'
  /api/v1/analytics/report/{reportType}:
    post:
      tags: [Analytics]
      summary: Generate a report
      description: |
        Generates and returns (or queues) a custom report.
        **Auth:** Bearer JWT — role: ORGANIZATION, ADMIN, or AUDITOR.
        **Rate limit:** 20 req / min per user.
      operationId: generateReport
      security:
        - bearerAuth: []
      parameters:
        - name: reportType
          in: path
          required: true
          schema:
            type: string
            enum: [campaign, donor, organization, platform, audit]
          example: campaign
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                entityId:
                  type: string
                  example: clxcamp001
                fromDate:
                  type: string
                  format: date-time
                toDate:
                  type: string
                  format: date-time
                format:
                  type: string
                  enum: [json, csv, pdf]
                  default: json
            example:
              entityId: clxcamp001
              fromDate: "2024-01-01T00:00:00Z"
              toDate: "2024-03-31T23:59:59Z"
              format: csv
      responses:
        "200":
          description: Report generated
          content:
            application/json:
              example:
                success: true
                data:
                  reportId: clxrpt001
                  status: COMPLETED
                  downloadUrl: https://storage.aidlink.org/reports/rpt001.csv
        "400":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "429":
          $ref: '#/components/responses/RateLimited'

  # ==========================================================================
  # SEARCH
  # ==========================================================================
...